requirements.txt
python-telegram-bot
openai
pyyaml
pydantic
pydantic-settings
python-dotenv
tiktoken
